[#164563148] Group tests
