+++ spec/support/shared_examples.rb
@@ -22,19 +22,7 @@ RSpec.shared_examples "user login" do |parameter|
   let(:user) { FactoryBot.create :facebook_user }
 
   before :each do
-    OmniAuth.config.test_mode = true
-
-    OmniAuth.config.mock_auth[:facebook] = OmniAuth::AuthHash.new({
-      "provider": "facebook",
-      "uid": user.identities.facebook.first.uid,
-      "email": user.email,
-      "name": user.name
-    })
-
-    visit new_user_session_path
-    click_link 'Login with Facebook'
-
-    wait_until { page.current_path == root_path }
+    login_as user, scope: :user
   end
 end
